Change 2
Comments have been made outlining what needs to be written/formatting of said page.

Only thing that needs to be done now is to structure the content separate of the website and figure out some way of hosting this (for free atleast)

Maybe put in a link to the UNC RELI and COMP SCI department?